Bugfix: switch to percpu reentrant protection
The prior solution was using per thread flags to denote whether or not we are
in memorizer. This could be problematic because in some cases we may be in an
IRQ context and thus are not executing on the old thread context. It likely
worked because 'current' was the last thread, but there could be gaps in this.
The new code uses a percpu flag to denote.

TODO: a pass needs to be made to evaluate where we need to make memorizer
non-reentrant. For example, if we are processing a memorizer write log: do we
need to disable interrupts for the whole function? Likely not. We only need to
make sure the wq_push is atomic. This would make the system faster. The real
issue with reentrance is when we have an allocation and then have to leave
memorizer to handle it. There can be a few cases:

- the allocator then reenters on the ld/st
- the ld/st instrumentation doesn't allocate any more so it doesn't need
  protection
- the allocator sleeps: I don't know what would happen here. If the allocator
  sleeps while we are *in-memorizer* then we will not allow any more entrances
  until that allocation finishes. So in this case I think we are still safe,
  however, we will be disabled on that CPU for any subsequent activities.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>