Move timestamp to global memorizer and fix lookup
The lookup table observes frees if a given allocation is overwriting a previous
one. This means that we've missed hooking the particular free. So we can track
it implicitly here, but need the timestamp for events to be global to all of
memorizer. Move it over to memorizer.h and employ it to mark frees detected in
this way.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>