Add reserve pool for kobjs and make alloc atomic
Once we enter into memorizer we cannot allow an interrupt to occur that may
allow control to go somewhere else and lead to missed tracing. So we need to
make it atomic by disabling interrupts. Additionally, because we might
operate in a no_sleep mode we also need to have a reserve memory pool in case
the kobj slab does not have any available objects. So add a reserve pool to
service such requests.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>