Hook induced frees stats
Want to track number of frees that occur while we are already "in-memorizer".
Add the stats. The interesting thing here is that these could be induced or
they could be frees that might occur if we did not disable interrupts properly.
As the code is it should only be induced, but this requires finer validation.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>