Change default behavior of memorizer to print live
Memorizer has the ability to printout all cmap entries or filter and print out
only free'd object cmap data. The point of this is to allow for iterative
printing during tracing, which would lead to double printing some object data
if the live object printing isn't left till the very end. The default of having
    this off might be confusing because a run might easily fit and thus someone
    print out cmap without considering the special case. So default to print
    out all and then provide documentation for how to print cmap while
    recording.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>