[wip] first guess at kasan stack check; no work
this gets the shadow region associated with a thing being written to.
The value returned is '0', which means that it is an okay access, but we
need to lookup the "type" of object here. I think that each object has a
guard, a 1/2 bit sequence associated with it. We need this round up.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>