Add general kobj: track access info
This extends the coverage of missing object access and most important
allows us to output data on what src PCs are doing with respect to
access. We may not have full information about what objects are being
modified, but we still can get a sense of where the worst offenders of
missed references are coming from. This data shows memcmp is a big
offender.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>