Use PC of called instruction to __asan_loadN
I previously noted that calls to __asan_loadN/__asan_storeN will not
allow us to track the access to a specific PC because the compiler
inserts these checks, however, they can account for a large number of
accesses. This modification uses the calling instruction as the
"accessing" instruction, but that is not the real instruction doing the
access.

It might be that this call happens immediate before the access and we
can use a simple heuristic, but for now, I wanted to make the coverage
issue complete. It means that our data analysis needs to take this into
account.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>