Use mempool allocator with emergency pool 4 access
The per access data structure stresses slab allocator, found out it is
just freezing and sometimes hitting a segfault. Switched to a mempool
allocator, which provides a bit more safety by always having around a
certain number of objects ready for use. Tweaking is needed but this
should be stable on work well.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>