[WIP]Create new mostly manual allocator 4 accesscounts
We non-deterministically getting errors on slab allocations. To fix this
I implemented a new custom allocator that builds on top of the page
allocator. The idea was to stress the Linux allocator less in general,
Although, the allocator appears to work a bit it deterministically hits
a deadlock, which surprisingly is impacted by the amount of RAM
available.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>