Memorizer: MEMBLOCK tracing and missing ref checks
This adds very simple support for memblock tracing. It does not deal
with overlapping blocks nor frees. It only tracks in a single list the
set of allocs then linearly searches this list if we miss on an access.
For the simple `ls` test we get a factor of 2 improvement on coverage:
60% to 80%. As more complex tests have shown better coverage this should
start to see us in the 90% range.

Signed-off-by: Nathan Dautenhahn <ndd@cis.upenn.edu>